added early stopping and choose best

+++ detector/tools/train_utils/train_utils.py
@@ -4,9 +4,11 @@ import os
 import torch
 import tqdm
 from torch.nn.utils import clip_grad_norm_
+# eval_loss用于评价验证集损失函数
+# eval_mAP用于评价验证集mAP
+from eval_utils.eval_utils import eval_loss
 
-
-def train_one_epoch(model, optimizer, train_loader, model_func, lr_scheduler, accumulated_iter, optim_cfg,
+def train_one_epoch(model, optimizer, train_loader,test_loader, model_func, lr_scheduler, accumulated_iter, optim_cfg,
                     rank, tbar, total_it_each_epoch, dataloader_iter, tb_log=None, leave_pbar=False):
     if total_it_each_epoch == len(train_loader):
         dataloader_iter = iter(train_loader)
@@ -35,14 +37,14 @@ def train_one_epoch(model, optimizer, train_loader, model_func, lr_scheduler, ac
         model.train()
         optimizer.zero_grad()
 
-        loss, tb_dict, disp_dict = model_func(model, batch)
+        train_loss, tb_dict, disp_dict = model_func(model, batch)
 
-        loss.backward()
+        train_loss.backward()
         clip_grad_norm_(model.parameters(), optim_cfg.GRAD_NORM_CLIP)
         optimizer.step()
 
         accumulated_iter += 1
-        disp_dict.update({'loss': loss.item(), 'lr': cur_lr})
+        disp_dict.update({'train_loss': train_loss.item(), 'lr': cur_lr})
 
         # log to console and tensorboard
         if rank == 0:
@@ -52,20 +54,22 @@ def train_one_epoch(model, optimizer, train_loader, model_func, lr_scheduler, ac
             tbar.refresh()
 
             if tb_log is not None:
-                tb_log.add_scalar('train/loss', loss, accumulated_iter)
+                tb_log.add_scalar('train/loss', train_loss, accumulated_iter)
                 tb_log.add_scalar('meta_data/learning_rate', cur_lr, accumulated_iter)
                 for key, val in tb_dict.items():
                     tb_log.add_scalar('train/' + key, val, accumulated_iter)
     if rank == 0:
         pbar.close()
-    return accumulated_iter
+    #val_loss=eval_loss(model,test_loader,model_func=model_func,accumulated_iter=accumulated_iter,rank=rank,tbar=tbar)
+    return accumulated_iter,train_loss
 
 
-def train_model(model, optimizer, train_loader, model_func, lr_scheduler, optim_cfg,
+def train_model(model, optimizer, train_loader, test_loader,model_func, lr_scheduler, optim_cfg,
                 start_epoch, total_epochs, start_iter, rank, tb_log, ckpt_save_dir, train_sampler=None,
                 lr_warmup_scheduler=None, ckpt_save_interval=1, max_ckpt_save_num=50,
-                merge_all_iters_to_one_epoch=False):
+                merge_all_iters_to_one_epoch=False,choose_best=True,use_early_stopping=True):
     accumulated_iter = start_iter
+    loss=float('inf')
     with tqdm.trange(start_epoch, total_epochs, desc='epochs', dynamic_ncols=True, leave=(rank == 0)) as tbar:
         total_it_each_epoch = len(train_loader)
         if merge_all_iters_to_one_epoch:
@@ -83,8 +87,8 @@ def train_model(model, optimizer, train_loader, model_func, lr_scheduler, optim_
                 cur_scheduler = lr_warmup_scheduler
             else:
                 cur_scheduler = lr_scheduler
-            accumulated_iter = train_one_epoch(
-                model, optimizer, train_loader, model_func,
+            accumulated_iter,train_loss_of_cur_epoch= train_one_epoch(
+                model, optimizer, train_loader,test_loader, model_func,
                 lr_scheduler=cur_scheduler,
                 accumulated_iter=accumulated_iter, optim_cfg=optim_cfg,
                 rank=rank, tbar=tbar, tb_log=tb_log,
@@ -93,21 +97,47 @@ def train_model(model, optimizer, train_loader, model_func, lr_scheduler, optim_
                 dataloader_iter=dataloader_iter
             )
 
-            # save trained model
-            trained_epoch = cur_epoch + 1
-            if trained_epoch % ckpt_save_interval == 0 and rank == 0:
+            # save model with choose best，选取最好的模型保存
+            if choose_best:
+                trained_epoch = cur_epoch + 1
+                best_ckpt_path=str(ckpt_save_dir / 'best')
+                last_ckpt_path=str(ckpt_save_dir / 'last')
+                # save best.pth
+                if train_loss_of_cur_epoch<loss:
+                    loss=train_loss_of_cur_epoch
+                    save_checkpoint(
+                            checkpoint_state(model, optimizer, trained_epoch, accumulated_iter), filename=best_ckpt_path,
+                        )
+                else:
+                    pass
+                # save last.pth
+                if trained_epoch==total_epochs:
+                    save_checkpoint(
+                            checkpoint_state(model, optimizer, trained_epoch, accumulated_iter), filename=last_ckpt_path,
+                        )
+                else:
+                    pass
+            else:
+            # save trained model without echoose best and early stopping
+                trained_epoch = cur_epoch + 1
+                if trained_epoch % ckpt_save_interval == 0 and rank == 0:
 
-                ckpt_list = glob.glob(str(ckpt_save_dir / 'checkpoint_epoch_*.pth'))
-                ckpt_list.sort(key=os.path.getmtime)
+                    ckpt_list = glob.glob(str(ckpt_save_dir / 'checkpoint_epoch_*.pth'))
+                    ckpt_list.sort(key=os.path.getmtime)
 
-                if ckpt_list.__len__() >= max_ckpt_save_num:
-                    for cur_file_idx in range(0, len(ckpt_list) - max_ckpt_save_num + 1):
-                        os.remove(ckpt_list[cur_file_idx])
+                    if ckpt_list.__len__() >= max_ckpt_save_num:
+                        for cur_file_idx in range(0, len(ckpt_list) - max_ckpt_save_num + 1):
+                            os.remove(ckpt_list[cur_file_idx])
 
-                ckpt_name = ckpt_save_dir / ('checkpoint_epoch_%d' % trained_epoch)
-                save_checkpoint(
-                    checkpoint_state(model, optimizer, trained_epoch, accumulated_iter), filename=ckpt_name,
-                )
+                    ckpt_name = ckpt_save_dir / ('checkpoint_epoch_%d' % trained_epoch)
+                    save_checkpoint(
+                        checkpoint_state(model, optimizer, trained_epoch, accumulated_iter), filename=ckpt_name,
+                    )
+            # TODO:write early stopping here
+            if use_early_stopping:
+                print("use early stopping")
+            else:
+                pass
 
 
 def model_state_to_cpu(model_state):
